Extracted the fleet names building logic
