+++ Service/FileContentBuilder.cs
@@ -19,12 +19,18 @@ namespace StellarisNameListGenerator.Service
 
         readonly INamesBuilder shipNamesBuilder;
         readonly INamesBuilder shipClassNamesBuilder;
+        readonly IFleetNamesBuilder fleetNamesBuilder;
         readonly INamesBuilder planetNamesBuilder;
 
-        public FileContentBuilder(INamesBuilder shipNamesBuilder, INamesBuilder shipClassNamesBuilder, INamesBuilder planetNamesBuilder)
+        public FileContentBuilder(
+            INamesBuilder shipNamesBuilder,
+            INamesBuilder shipClassNamesBuilder,
+            IFleetNamesBuilder fleetNamesBuilder,
+            INamesBuilder planetNamesBuilder)
         {
             this.shipNamesBuilder = shipNamesBuilder;
             this.shipClassNamesBuilder = shipClassNamesBuilder;
+            this.fleetNamesBuilder = fleetNamesBuilder;
             this.planetNamesBuilder = planetNamesBuilder;
         }
 
@@ -36,7 +42,7 @@ namespace StellarisNameListGenerator.Service
             content += $"### {nameList.Name}{Environment.NewLine}";
             content += $"### Leaders: {GetRandomLeaderName(nameList)}, {GetRandomLeaderName(nameList)}{Environment.NewLine}";
             content += $"### Ships: {GetRandomShipName(nameList)}, {GetRandomShipName(nameList)}{Environment.NewLine}";
-            content += $"### Fleets: {GetRandomFleetName(nameList)}, {GetRandomFleetName(nameList)}{Environment.NewLine}";
+            content += $"### Fleets: {fleetNamesBuilder.GetRandomName(nameList)}, {fleetNamesBuilder.GetRandomName(nameList)}{Environment.NewLine}";
             content += $"### Colonies: {GetRandomPlanetName(nameList)}, {GetRandomPlanetName(nameList)}{Environment.NewLine}";
             content += Environment.NewLine;
 
@@ -80,7 +86,7 @@ namespace StellarisNameListGenerator.Service
                         break;
 
                     case "FleetNames":
-                        sections[section.Key] = BuildFleetNames(nameList);
+                        sections[section.Key] = fleetNamesBuilder.Build(nameList);
                         break;
 
                     case "ArmyNames":
@@ -115,17 +121,6 @@ namespace StellarisNameListGenerator.Service
 
             return $"{randomisableLine}{Environment.NewLine}";
         }
-
-        string BuildFleetNames(NameList nameList)
-        {
-            string content = string.Empty;
-
-            IEnumerable<NameGroup> fleetNames = GenerateFleetNames(nameList);
-            content += BuildNameArray(fleetNames, "fleet_names", 1, nameList.Armies.FleetSequentialName);
-
-            return content;
-        }
-
         string BuildArmyNames(NameList nameList)
         {
             string content = string.Empty;
@@ -406,35 +401,6 @@ namespace StellarisNameListGenerator.Service
             return string.Empty.PadRight(levels * IndentationSize, ' ');
         }
 
-        IEnumerable<NameGroup> GenerateFleetNames(NameList nameList)
-        {
-            return nameList.Armies.Fleet
-                .Concat(GenerateFleetNamesCategory(nameList, "Armadas", "The {0} Armada"))
-                .Concat(GenerateFleetNamesCategory(nameList, "Battle Groups", "The {0} Battle Group"))
-                .Concat(GenerateFleetNamesCategory(nameList, "Corps", "The {0} Corps"))
-                .Concat(GenerateFleetNamesCategory(nameList, "Expeditionary Fleets", "The {0} Expeditionary Fleet"))
-                .Concat(GenerateFleetNamesCategory(nameList, "Fleets", "The {0} Fleet"))
-                .Concat(GenerateFleetNamesCategory(nameList, "Flotillas", "The {0} Flotilla"))
-                .Concat(GenerateFleetNamesCategory(nameList, "Squadrons", "{0} Squadron"))
-                .Concat(GenerateFleetNamesCategory(nameList, "Starfleets", "{0} Starfleet"))
-                .Concat(GenerateFleetNamesCategory(nameList, "Strike Forces", "Strike Force {0}"))
-                .Concat(GenerateFleetNamesCategory(nameList, "Strike Teams", "Strike Teams {0}"))
-                .Concat(GenerateFleetNamesCategory(nameList, "Task Forces", "Task Force {0}"))
-                .ToList();
-        }
-
-        IEnumerable<NameGroup> GenerateFleetNamesCategory(NameList nameList, string category, string nameFormat)
-        {
-            IList<NameGroup> fleetNames = new List<NameGroup>();
-            
-            fleetNames.Add(GenerateUnifiedNameGroup(nameList.Warfare.Weapons.All, category, "Weapons", nameFormat));
-            fleetNames.Add(GenerateUnifiedNameGroup(nameList.Warfare.MilitaryUnitTypes, category, "Military Unit Types", nameFormat));
-            fleetNames.Add(GenerateUnifiedNameGroup(nameList.Warfare.ShipTypes, category, "Ship Types", nameFormat));
-            fleetNames.Add(GenerateUnifiedNameGroup(nameList.BiosphereNames.MythologicalCreatures, category, "Mythological Creatures", nameFormat));
-
-            return fleetNames;
-        }
-
         NameGroup GenerateUnifiedNameGroup(IEnumerable<NameGroup> nameGroups, string category, string groupName, string nameFormat)
         {
             NameGroup group = new NameGroup();
@@ -551,27 +517,6 @@ namespace StellarisNameListGenerator.Service
                 .Concat(nameList.Warfare.Weapons.All);
 
             return shipNames.SelectMany(x => x.Values).GetRandomElement(random);
-        }
-
-        string GetRandomFleetName(NameList nameList)
-        {
-            IEnumerable<NameGroup> fleetNames = GenerateFleetNames(nameList);
-
-            if (fleetNames.Any(x => !x.IsEmpty))
-            {
-                return fleetNames
-                    .SelectMany(x => x.Values)
-                    .GetRandomElement(random);
-            }
-
-            List<string> cardinalNumbers = new List<string> { "1", "10", "33", "42", "56", "86", "101", "303", "500", "613", "743", "873" };
-            List<string> ordinalNumbers = new List<string> { "1st", "21st", "101st", "42nd", "62nd", "72nd", "53rd", "83rd", "123rd", "103rd", "4th", "12th", "14th", "404th" };
-            List<string> romanNumbers = new List<string> { "I", "II", "IV", "XI", "XXXII", "CXXXII", "CDII", "DLXII" };
-
-            return nameList.Armies.FleetSequentialName
-                .Replace("%O%", ordinalNumbers.GetRandomElement())
-                .Replace("%C%", cardinalNumbers.GetRandomElement())
-                .Replace("%R%", romanNumbers.GetRandomElement());
         }
 
         string GetRandomPlanetName(NameList nameList)
